Document Logger.Backends.Console
